📜Finally added descriptions
Also added qprogram.removegates() and qprogram.cleargates() to make gate manipulation easier.
Fixed minor bugs.
Almost done with the first version, just a few minor things left before merging with main.
