+++ quantum.py
@@ -1,6 +1,6 @@
 from math import atan, pi, log, cos, sin
 from random import random
-from typing import Type
+from typing import List
 from copy import deepcopy
 mple = True
 
@@ -12,9 +12,19 @@ except ModuleNotFoundError:
 
 
 PI = pi
+Pi = pi
+
 
 class comp(object):
+    """
+    Class for dealing with complex numbers.
+    """
     def __init__(self, a : float = 0, b : float = 0):
+        """
+        Defines a complex number z = a + ib
+        To define a complex number in polar form, use comp.polar().
+        If `a` is of type comp, it simply defines a new identical complex number.
+        """
         if type(a) == comp:
             self.a, self.b = a.a, a.b
         else:
@@ -26,8 +36,11 @@ class comp(object):
         except ZeroDivisionError:
             self.theta = pi/2 * (1 if self.b >= 0 else -1)
 
-    #theta is measured in radians; use quantum.pi for π
     def polar(r : float, theta : float):
+        """
+        Returns a comp in the given polar coordinates.
+        Theta is measured in radians; use quantum.pi or Pi or PI for π
+        """
         return comp(r * cos(theta), r * sin(theta))
 
 
@@ -48,6 +61,10 @@ class comp(object):
         return (self + (other * -1))
 
     def getcomplex(number : object or float or int):
+        """
+        Returns the comp form of whatever is passed in as input.
+        If input is comp, the SAME object is returned (as opposed to comp(z), which gives a copy).
+        """
         if type(number) == comp: return number
         return comp(number)
 
@@ -71,13 +88,26 @@ class comp(object):
             return comp(self.a * other, self.b * other)
     
     def inverse(self):
+        """
+        Returns (1/z).
+        """
         return comp( self.a/pow(self.r, 2) , (-self.b)/pow(self.r, 2))
     
     def conjugate(self):
+        """
+        Returns z̄.
+        """
         conj = comp(self.a, (-1 * self.b))
         return conj
 
 class Matrix(object):
+    """
+    Matrix of 2 dimensions with entries of type comp or float or int.
+    A matrix on its own is a valid quantum gate that can be added to a qprogram via qprogram.addgates().
+
+    Matrix.shape is the order of the matrix.
+    Matrix.span is number of qbits the gate is supposed to input and output.
+    """
 
     def __eq__(self, o: object) -> bool:
         if (self.nrows != o.nrows) or (self.ncols != o.ncols): return False
@@ -87,7 +117,17 @@ class Matrix(object):
         return True
 
     def __init__(self, r : int, c : int = -1, id : str = None):
+        """
+        Matrix(integer) returns a integer x integer Identity matrix.
+        
+        Matrix(a, b) returns a matrix of order a x b.
+        
+        Matrix(List[List[float or int or comp]]) converts a 2D matrix from list form to Matrix type.
+
+        `id` is the gateid of the quantum gate (preferably 1 or a few chars at most).
+        """
         if type(r) == list:
+            assert r != [], "Empty list cannot be converted to a matrix."
             mat = r
             self.nrows = len(mat)
             self.ncols = len(mat[0])
@@ -138,7 +178,10 @@ class Matrix(object):
         if type(other) == comp: return(self * other.inverse())
         return (self * (1/other))
     
-    def determinant(matrix):
+    def determinant(matrix) -> float or comp:
+        """
+        Returns the determinant of this matrix.
+        """
         assert matrix.nrows == matrix.ncols, f"Matrix must a square matrix; got dimensions {matrix.nrows}x{matrix.ncols}"
         if matrix.ncols == 1: return matrix.rows[0][0]
         if matrix.ncols == 2:
@@ -155,6 +198,9 @@ class Matrix(object):
         return det
 
     def adjoint(matrix):
+        """
+        Returns the adjoint of this matrix.
+        """
         assert matrix.nrows == matrix.ncols, f"Matrix must a square matrix; got dimensions {matrix.nrows}x{matrix.ncols}"
         adj = deepcopy(matrix)
         if matrix.nrows == 2:
@@ -176,6 +222,9 @@ class Matrix(object):
         return adj
     
     def inverse(self):
+        """
+        Returns the inverse of this matrix.
+        """
         return (self.adjoint() / self.determinant())
 
 
@@ -186,14 +235,19 @@ class Matrix(object):
         return string.strip()
     
     def transpose(self):
+        """
+        Returns the transpose of this matrix.
+        """
         trans = Matrix(self.ncols, self.nrows)
         for r in range(trans.nrows):
             for c in range(trans.ncols):
                 trans.rows[r][c] = self.rows[c][r]
         return trans
     
-    #this is elementwise conjugate of the matrix;
     def conjugate(self):
+        """
+        Returns the elementwise conjugate of the matrix.
+        """
         conj = Matrix(self.nrows, self.ncols)
         for r in range(conj.nrows):
             for c in range(conj.ncols):
@@ -201,24 +255,41 @@ class Matrix(object):
         return conj
     
     def dagger(matrix):
+        """
+        Returns the conjugate of the matrix.
+        """
         return matrix.conjugate().transpose()
     
     def getlist(self) -> list:
+        """
+        If the matrix is a column or row matrix, returns a list form of the matrix.
+        """
         if self.shape[0] == 1: return self.rows[0]
         elif self.shape[1] == 1: return self.transpose().rows[0]
         else: print(f"Cannot meaningfully convert to list; at least 1 dimension must be 1; shape : {self.shape}")
 
-def mod(x, squared = False):
+def mod(x, squared = False) -> float:
+    """
+    Returns the mod value of x.
+    If squared is True, returns the square of the mod value.
+    """
     if type(x) == type(comp(0, 0)):
         return pow(x.r, 2) if squared else x.r
     else:
         return (x if x >= 0 else (-1 * x)) if not squared else pow(x, 2)
 
-def qbit(bit : int = 0) -> list:
+def qbit(bit : int = 0) -> List[int]:
+    """
+    Returns a Qubit set to value bit.
+    Qubit 0 is [1, 0], and 1 is [0, 1].
+    """
     return [0, 1] if bit else [1, 0]
 
 
-def qtensor(q1 : list, q2 : list) -> list:
+def qtensor(q1 : List[int or float or comp], q2 : List[int or float or comp]) -> List[int or float or comp]:
+    """
+    Returns the tensor product Q1⛒Q2.
+    """
     tensorproduct = []
     for each in q1:
         for every in q2:
@@ -227,6 +298,9 @@ def qtensor(q1 : list, q2 : list) -> list:
 
 
 def mtensor(m1 : Matrix, m2 : Matrix) -> Matrix:
+    """
+    Returns the Matrix tensor product M1⛒M2.
+    """
     r = m1.nrows * m2.nrows
     c = m1.ncols * m2.ncols
     product = Matrix(r, c)
@@ -238,7 +312,12 @@ def mtensor(m1 : Matrix, m2 : Matrix) -> Matrix:
                     product.rows[(m2.nrows * each) + x][(m2.ncols * i) + y] = comp(m1.rows[each][i]) * comp(m2.rows[x][y])
     return product
 
-def MEASURE(tensor: list) -> list:
+def MEASURE(tensor: List[int or float or comp]) -> List[int]:
+    """
+    Performs a measurement on a system of qubits.
+    `tensor` is a tensor in list form that represents the state of the system.
+    Returns a sparse list with result of measurement.
+    """
     prev = 0
     ranges = list( map( lambda x: mod(x, True), tensor ) )
     m = random()
@@ -252,6 +331,9 @@ def MEASURE(tensor: list) -> list:
     return res
 
 def printreal(matrix : Matrix) -> None:
+    """
+    Prints the matrix array with only real parts of the entries.
+    """
     thing = ""
     for each in matrix.rows:
         line = ""
@@ -261,6 +343,9 @@ def printreal(matrix : Matrix) -> None:
     print(thing)
 
 def printimg(matrix : Matrix) -> None:
+    """
+    Prints the matrix array with only imaginary parts of the entries.
+    """
     thing = ""
     for each in matrix.rows:
         line = ""
@@ -271,6 +356,12 @@ def printimg(matrix : Matrix) -> None:
 
 
 def plotmeasurement(measurement : list, binary = True, name : str = None) -> None:
+    """
+    Plots the measurement result graphically.
+    `name` is the title for the graph.
+    `binary` controls if the states are represented in binary or decimal.
+    If matplotlib has not been installed, the terminal will automatically be used for plotting.
+    """
     eigenvectors = [('Ψ' + (('0' * (int(log(len(measurement), 2)) - len(str(bin(i))[2:]))) + str(bin(i))[2:])) for i in range(len(measurement))]
     if not binary : eigenvectors = [('Ψ' + str(i)) for i in range(len(measurement))]
     if mple:
@@ -284,6 +375,11 @@ def plotmeasurement(measurement : list, binary = True, name : str = None) -> Non
         plotinterminal(measurement=measurement, binary= binary)
 
 def plotinterminal(measurement : list, binary = True, name : str = None) -> None:
+    """
+    Plots the measurement result graphically in terminal.
+    `name` is the title for the graph.
+    `binary` controls if the states are represented in binary or decimal.
+    """
     eigenvectors = [('Ψ' + (('0' * (int(log(len(measurement), 2)) - len(str(bin(i))[2:]))) + str(bin(i))[2:])) for i in range(len(measurement))]
     if not binary : eigenvectors = [('Ψ' + str(i)) for i in range(len(measurement))]
     graph = f"\n{'' if name is None else name}"
@@ -296,7 +392,13 @@ def plotinterminal(measurement : list, binary = True, name : str = None) -> None
     print(graph)
     return graph
 
-def run(shots : int, state : list, binary : bool = True, graph : bool = False, terminal : bool = False, name : str = None) -> None:
+def run(shots : int, state : List[int or float or comp], binary : bool = True, graph : bool = False, terminal : bool = False, name : str = None) -> None:
+    """
+    Runs measurements on a state of the system.
+    The state is measured `shots` number of times, and percentages of each outcome is recorded.
+    If `graph` is set to True, the result is also plotted graphically.
+    """
+    
     sl = int(log(len(state), 2))
     res = [0 for i in range(len(state))]
     for each in range(shots):
@@ -314,7 +416,11 @@ def run(shots : int, state : list, binary : bool = True, graph : bool = False, t
         if not terminal: plotmeasurement([((each/shots) * 100) for each in res], binary = binary, name = name)
         else: plotinterminal([((each/shots) * 100) for each in res], binary = binary, name = name)
 
-def extract(measurement : list, qbitindex : int) -> list:
+def extract(measurement : List[int], qbitindex : int) -> List[int]:
+    """
+    Returns the state of Qubit at index `qbitindex` in `measurement`.
+    NOTE: returns qubits in list form ([0, 1] or [1, 0]).
+    """
     nqbits = int(log(len(measurement), 2))
     index = bin(measurement.index(1))[2:]
     index = ('0' * (nqbits - len(index))) + index
@@ -323,6 +429,7 @@ def extract(measurement : list, qbitindex : int) -> list:
 
 def validgate(gate : Matrix) -> bool:
     """
+    Verifies if a Matrix object is a valid quantum gate.
     This aint working for some reason dont use it
     """
     print((gate.nrows == gate.ncols))
@@ -332,191 +439,85 @@ def validgate(gate : Matrix) -> bool:
     )
 
 
-hgate = Matrix([
+HGATE = Matrix([
     [comp(1),  comp(1)],
     [comp(1), comp(-1)]
 ], id = 'h') * pow(2, -0.5)
 
-igate = Matrix(2, id = 'i')
+IGATE = Matrix(2, id = 'i')
 
-cnot0 = Matrix(4, id = 'c0')
-cnot0.rows[2], cnot0.rows[3] = cnot0.rows[3], cnot0.rows[2]
+CNOT0 = Matrix(4, id = 'c0')
+CNOT0.rows[2], CNOT0.rows[3] = CNOT0.rows[3], CNOT0.rows[2]
 
-cnot1 = Matrix(4, id = 'c1')
-cnot1.rows[1], cnot1.rows[3] = cnot1.rows[3], cnot1.rows[1]
+CNOT1 = Matrix(4, id = 'c1')
+CNOT1.rows[1], CNOT1.rows[3] = CNOT1.rows[3], CNOT1.rows[1]
 
-xgate = Matrix([
+XGATE = Matrix([
     [comp(), comp(1)],
     [comp(1), comp()]
 ], id = 'x')
 
-ygate = Matrix([
+YGATE = Matrix([
     [comp(), comp(0, -1)],
     [comp(0, 1), comp( )]
 ], id = 'y')
 
-zgate = Matrix([
+ZGATE = Matrix([
     [comp(1), comp( )],
     [comp(), comp(-1)]
 ], id = 'z')
 
-tgate = Matrix([
+TGATE = Matrix([
     [comp(1), comp()],
     [comp(), comp.polar(1, pi/4)]
 ], id = 't')
 
-swapgate = Matrix(4, id = 'swap')
-swapgate.rows[1], swapgate.rows[2] = swapgate.rows[2], swapgate.rows[1]
+SWAPGATE = Matrix(4, id = 'swap')
+SWAPGATE.rows[1], SWAPGATE.rows[2] = SWAPGATE.rows[2], SWAPGATE.rows[1]
 
-mtensoridentity = Matrix([
+MTENSORIDENTITY = Matrix([
     [comp(1)]
 ], id = 'm')
 
-toffoligate = Matrix(8, id = 'tof')
-toffoligate.rows[6], toffoligate.rows[7] = toffoligate.rows[7], toffoligate.rows[6]
+TOFFOLIGATE = Matrix(8, id = 'tof')
+TOFFOLIGATE.rows[6], TOFFOLIGATE.rows[7] = TOFFOLIGATE.rows[7], TOFFOLIGATE.rows[6]
 
-def rgate(angle : float) -> Matrix:
+def RGATE(angle : float) -> Matrix:
+    """
+    Returns a matrix of the form:
+
+    [ sin(angle)   cos(angle) ]
+    [ cos(angle)  -sin(angle) ]
+
+    `angle` is in radians.
+    """
     matrix = Matrix(2)
     matrix.rows[0][0] = comp(sin(angle), 0)
     matrix.rows[0][1] = comp(cos(angle), 0)
     matrix.rows[1][0] = comp(cos(angle), 0)
     matrix.rows[1][1] = comp(-sin(angle), 0)
-    matrix.gateid = f'r({angle})'
+    matrix.gateid = f'r({round(angle, 2)})'
     return matrix
-
-def legacy_HGATE(m : int = 1) -> Matrix:
-    if m == 0: return 1
-    inner = legacy_HGATE(m-1)
-    matrix = Matrix(int(pow(2, m)), int(pow(2, m)))
-    matrix.gateid = 'h'
-    if m == 1:
-        matrix = matrix * (1/pow(2, 0.5))
-        matrix.rows[-1][-1] =  matrix.rows[-1][-1] * -1
-        matrix.gateid = 'h'
-        return matrix
-    for each in range(int(pow(2, m))):
-        for i in range(int(pow(2, m))):
-            matrix.rows[each][i] = inner.rows[each % int(pow(2, m-1))][i % int(pow(2, m-1))]
-            if (pow(2, m-1) <= each) and (pow(2, m-1) <= i): matrix.rows[each][i] *= -1
-    matrix = matrix * (1/pow(2, 0.5))
-    matrix.gateid = 'h'
-    return matrix
-
-def legacy_NGATE(n : int = 1) -> Matrix:
-    gate = Matrix(pow(2, n), pow(2, n))
-    gate = gate * 0
-    # buffer = Matrix(pow(2, n), pow(2, n))
-    for each in range(pow(2, n)):
-        for i in range(pow(2, n)):
-            if (each + i) == (pow(2, n) - 1): gate.rows[each][i] = comp(1, 0)
-    gate.gateid = 'n'
-    return gate
-
-def legacy_IGATE(n : int = 1) -> Matrix:
-    gate = Matrix(pow(2, n), pow(2, n))
-    gate = gate * 0
-    for each in range(pow(2, n)):
-        for i in range(pow(2, n)):
-            if each == i:
-                gate.rows[each][i] = comp(1, 0)
-    gate.gateid = 'i'
-    return gate
-
-def legacy_SETTOGATE(value : int, n : int = 1):
-    if n == 1:
-        matrix = Matrix(2, 2) * 0
-        matrix.rows[value][0] = comp(1, 0)
-        matrix.rows[value][1] = comp(1, 0)
-        return matrix
-    else:
-        return mtensor(legacy_SETTOGATE(value = value), legacy_SETTOGATE(value = value, n =  n - 1))
-
-def legacy_NOT(qbit: list) -> list:
-    return (legacy_NGATE(int(log(len(qbit), 2))) ** qbit)
-
-def legacy_IDEN(qbit : list) -> list:
-    return qbit
-
-def legacy_HAD(qbit : list) -> list:
-    # print(qbit)
-    return legacy_HGATE(int(log(len(qbit), 2))) ** qbit
-
-def legacy_FLIP(state: list) -> list:
-    flipped = []
-    for each in state:
-        each = comp.getcomplex(each)
-        each.b *= -1
-        flipped.append(each)
-    return flipped
-
-def legacy_flipgate(gate : Matrix) -> Matrix:
-    flipped = Matrix(gate.nrows, gate.ncols)
-    flipped.gateid = gate.gateid
-    flipped.rows = deepcopy(gate.rows)
-    flipped.rows = flipped.rows[::-1]
-    return flipped
-
-def legacy_SHIFTGATE(state : list, phase : float) -> list:
-    shifted = []
-    rotor = comp.polar(1, phase)
-    for each in state:
-        shifted.append(comp.getcomplex(each) * rotor)
-    return shifted
-
-def legacy_SHIFT(phase : float) -> Type[lambda x: x]:
-    return lambda x: legacy_SHIFTGATE(x, phase)
-
-
-def legacy_RGATE(angle : float):
-    matrix = Matrix(2, 2)
-    matrix.rows[0][0] = comp(sin(angle), 0)
-    matrix.rows[0][1] = comp(cos(angle), 0)
-    matrix.rows[1][0] = comp(cos(angle), 0)
-    matrix.rows[1][1] = comp(-sin(angle), 0)
-    matrix.gateid = 'r'
-    return matrix
-
-
-def legacy_XGATE() -> Matrix:
-    matrix = Matrix(2, 2)
-    matrix.rows[0][0] = 0
-    matrix.rows[1][1] = 0
-    matrix.gateid = 'x'
-    return matrix
-
-def legacy_YGATE() -> Matrix:
-    matrix = Matrix(2, 2) * comp(0, 1)
-    matrix.rows[0][0] = 0
-    matrix.rows[1][1] = 0
-    matrix.rows[0][1] *= -1
-    matrix.gateid = 'y'
-    return matrix
-
-def legacy_ZGATE() -> Matrix:
-    matrix = legacy_IGATE()
-    matrix.rows[1][1] *= -1
-    matrix.gateid = 'z'
-    return matrix
-
-
-def legacy_PHASEGATE(phase : float) -> Matrix:
-    matrixgate = Matrix(2, 2) * 0
-    matrixgate.rows[0][0] = 1
-    matrixgate.rows[1][1] = comp.polar(1, phase)
-    matrixgate.gateid = 'phase'
-    return matrixgate
-
-
-
 
 
 class basischange:
+    """
+    A class for dealing with basis change.
+    Instance of this object can be passed into qprogram for changing the basis.
+    """
     def __init__(self, transformationmatrix : Matrix) -> None:
+        """
+        Transformation matrix is the matrix for changing basis of the system.
+        Must be a 2x2 matrix.
+        """
         r, c = transformationmatrix.nrows, transformationmatrix.ncols
         assert (r == 2) and (c == 2) , f"transformationmatrix must be a 2x2 matrix; got {r}x{c}"
         self.transmat = transformationmatrix
     
     def get(self, nqubitsinsystem : int = 1):
+        """
+        Returns the transformation matrix for an n-qubit system.
+        """
         if nqubitsinsystem == 1: return self.transmat
         else:
             retter = deepcopy(self.transmat)
@@ -524,98 +525,33 @@ class basischange:
                 retter = mtensor(self.transmat, retter)
             return retter
 
-
-
-
-
-def legacy_CNOTGATEOLD(controlindex : int = 0, targetindex : int = 1) -> Matrix:
-    cg = Matrix(4, 4)
-    cg = cg * 0
-    for each in range(4):
-        for i in range(4):
-            if each == i: cg.rows[each][i] = comp(1, 0)
-    if controlindex == 0 and targetindex == 1:
-        buffer = cg.rows[-1]
-        cg.rows[-1] = cg.rows[-2]
-        cg.rows[-2] = buffer
-    elif controlindex == 1 and targetindex == 0:
-        buffer = cg.rows[-1]
-        cg.rows[-1] = cg.rows[1]
-        cg.rows[1] = buffer
-    cg.gateid = 'cnot'
-    return cg
-
-def legacy_CNOTGATE(nqbits : int = 2, controlindex : int = 0, targetindex : int = 1) -> Matrix:
-    if controlindex == targetindex: raise BaseException("control and target cannot be the same")
-    if nqbits <= 1: raise BaseException("nqbits cannot be less than 2")
-    if nqbits == 2: return legacy_CNOTGATEOLD(controlindex = controlindex, targetindex = targetindex)
-    t = mtensor(legacy_CNOTGATEOLD(controlindex = controlindex, targetindex = targetindex), legacy_CNOTGATE(nqbits - 1, controlindex= controlindex, targetindex=targetindex))
-    t.gateid = 'cnot' + str(controlindex)
-    return t
-
-def legacy_CNOTR(controlbitindex : int = 0) -> Matrix:
-    gate = Matrix(4, 4) * 0
-    gate.gateid = 'cnotr'
-
-    gate.rows[0][0] = comp(1, 0)
-    gate.rows[1][1] = comp(1, 0)
-    gate.rows[2][2] = comp(1, 0)
-    gate.rows[3][3] = comp(1, 0)
-
-    if controlbitindex == 0: gate.rows[2], gate.rows[3] = gate.rows[3], gate.rows[2]
-    else: gate.rows[1], gate.rows[3] = gate.rows[3], gate.rows[1]
-
-    return gate
-
-def legacy_FLIPPEDCNOTGATE(nqbits : int = 2) -> Matrix:
-    fcnot = Matrix(4, 4) * 0
-    fcnot.rows[0][0] = comp(1, 0)
-    fcnot.rows[1][3] = comp(1, 0)
-    fcnot.rows[2][2] = comp(1, 0)
-    fcnot.rows[3][1] = comp(1, 0)
-    fcnot.gateid = 'fcnot'
-    return fcnot
-
-
-def legacy_CNOT(qcontrol : list, qtarget : list) ->list:
-    tens = qtensor(qcontrol, qtarget)
-    result = legacy_CNOTGATE() ** tens
-    return result
-
-
-
-
-
-
-
-
-legacy_reprs = {
-    'r' : '[ R ]',
-    'i' : '-----',
-    'n' : '[ ~ ]',
-    'x' : '[ X ]',
-    'y' : '[ Y ]',
-    'z' : '[ Z ]',
-    'h' : '[ H ]',
-    'cnot' : '[ ⛒ ]',
-    'fcnot' : '[ ⛒ ]',
-    'phase' : '[ θ ]',
-    'pin' : '[ ☉ ]',
-    'c0' : '[ 0 ]',
-    'c1' : '[ 1 ]',
-    'm' :  '-----',
-    's' : '[ s ]'
-}
-
-
-def legacy_getrepr(gate : Matrix) -> str:
-    try : return legacy_reprs[gate.gateid]
-    except : 
-        print(f"Gate represntation not found: {gate.gateid}")
-        return '[ ! ]'
+class Block:
+    pass
 
 class qprogram(object):
+    """
+    Class for a Quantum Program.
+    This class deals in the gates and circuits model of quantum computing.
+
+    The normal pipeline for running a qprogram is to instantiate it, add gates, compile and then run.
+    
+    Example:
+    `
+    entangler = qprogram(2, name = 'quantum entanglement')
+    entangler.addgates(0, [HGATE, CNOT0])
+    entangler.compile()
+    entangler.run(graph = True)
+    `
+
+    NOTE: It is assumed that measurement takes place at the end of the qprogram in the circuit diagram.
+    """
     def __init__(self, nqbits : int, custom : list = None, bchange : basischange = None, name : str = None) -> None:
+        """
+        `nqbits` -> no. of qubits in the system.
+        `custom` -> a list of integers that sets the initial state of each of the qubits (by default 0 for all).
+        `bchange` -> a BasisChange object for changing the basis.
+        `name` -> name of the Quantum program.
+        """
         self.qbits = []
         self.nqbits = nqbits
         self.cache = None
@@ -630,45 +566,42 @@ class qprogram(object):
         self.programmat = None
 
     def addgates(self, qbitindex : int, gates : list):
+        """
+        Adds gates to the qprogram.
+
+        `qbitindex` is the index of the bit to which gates are added.
+        `gates` is a list containing the sequence of gates to be added.
+        """
         self.cache = None
         self.programmat = None
         self.gates[qbitindex] += gates
+    
+    def removegate(self, qbitindex : int, gateindex : int = -1) -> None:
+        """
+        Removes the gate at the specified position.
+        
+        If gate index is not provided, by default the last gate in the sequence is removed.
+        """
+        del self.gates[qbitindex][gateindex]
+
+    def cleargates(self, qbitindex : int) -> None:
+        """
+        Removes all gates on the circuit line corresponding to Qubit at `qbitindex`.
+        """
+        self.gates[qbitindex] = []
 
     def __repr__(self) -> str:
         if self.repr is None : return f"The program {'' if self.name is None else self.name} is yet to be compiled"
         return self.repr
-
-    def legacycompile(self, verbose : bool = False, showcompilationresult : bool = True):
-        if showcompilationresult : print(f"\nCompiling {'program' if self.name is None else self.name}...")
-        longest = 0
-        for each in range(len(self.gates)):
-            for i in range(len(self.gates[each])):
-                if self.gates[each][i].gateid == 'cnot':
-                    pin = legacy_IGATE()
-                    pin.gateid = 'pin'
-                    self.gates[each - 1].insert(i, pin)
-                elif self.gates[each][i].gateid == 'fcnot':
-                    pin = legacy_IGATE()
-                    pin.gateid = 'pin'
-                    self.gates[each + 1].insert(i, pin)
-        for each in range(len(self.gates)):
-            if longest < len(self.gates[each]): longest = len(self.gates[each])
-        
-        for each in range(len(self.gates)):
-            self.gates[each] += [legacy_IGATE() for i in range(longest - len(self.gates[each]))]
-        
-        self.calcrepr()
-        
-        if showcompilationresult:
-            print(self)
-            print("Compilation complete!\n")
-
-        if verbose and self.cache is not None: print(self.cache)
-
-        self.cache = None
-        return self
     
     def compile(self, verbose : bool = False, showcompilationresult : bool = True):
+        """
+        Compiles the qprogram. This step is necessary before running the qprogram, and everytime the circuit-gate diagram is changed.
+
+        If `verbose`, debug info is showed as well.
+        If `showcompilationresult`, the final block diagram of the program is displayed.
+        """
+
         if showcompilationresult : print(f"\nCompiling {'program' if self.name is None else self.name}...")
 
         for each in self.gates:
@@ -680,14 +613,14 @@ class qprogram(object):
             if len(each) > longest : longest = len(each)
         if verbose : print("longest set of gates :",longest)
         for each in range(len(self.gates)):
-            self.gates[each] += [igate for _ in range(longest - len(self.gates[each]))]
+            self.gates[each] += [IGATE for _ in range(longest - len(self.gates[each]))]
 
         for each in range(len(self.gates)):
             for i in range(len(self.gates[each])):
                 try:
                     for x in range(self.gates[each][i].span - 1):
                         x += 1
-                        if self.gates[each + x][i].gateid != 'm': self.gates[each + x].insert(i, mtensoridentity)
+                        if self.gates[each + x][i].gateid != 'm': self.gates[each + x].insert(i, MTENSORIDENTITY)
                 except IndexError:
                     print(f"Invalid position ({each}, {i}) for gate ({self.gates[each][i].gateid}) of span {self.gates[each][i].span}\nAborting compilation...")
                     return
@@ -703,7 +636,7 @@ class qprogram(object):
             if len(each) > longest : longest = len(each)
         if verbose : print("longest set of gates :",longest)
         for each in range(len(self.gates)):
-            self.gates[each] += [igate for _ in range(longest - len(self.gates[each]))]        
+            self.gates[each] += [IGATE for _ in range(longest - len(self.gates[each]))]        
         
         if verbose:
             print()
@@ -720,7 +653,7 @@ class qprogram(object):
             if allidentities:
                 if verbose: print(f"Skipping tensor product at {each} due to all I gates...")
                 continue
-            prod = mtensoridentity
+            prod = MTENSORIDENTITY
             for i in range(len(self.gates)):
                 prod = mtensor(prod, self.gates[i][each])
             archways.append(prod)
@@ -745,18 +678,12 @@ class qprogram(object):
             print(self.repr)
             print(f"\nCompilation{(' of ' + self.name) if self.name is not None else ''} complete!\n")
 
-    def legacycalcrepr(self):
-        string = f"\n{'' if self.name is None else self.name}"
-        if self.name is not None: string += '\n'
-        for each in range(self.nqbits):
-            line = f"q{str(each)} ({self.qbits[each][1]}) ⮕ ---"
-            for i in range(len(self.gates[each])):
-                line += legacy_getrepr(self.gates[each][i])
-                line += "---"
-            string += line + '\n'
-        self.repr = string
+    
     
     def calcrepr(self):
+        """
+        Makes and returns the string representation of the block diagram of the qprogram.
+        """
         string = f"\n{'' if self.name is None else self.name}\n"
 
         reprs = [f'q{i}({self.qbits[i].index(1)}) ⮕  ---' for i in range(len(self.gates))]
@@ -796,67 +723,66 @@ class qprogram(object):
             string += '---\n'
         self.repr = string
     
-    def getstate(self, verbose : bool = False, usecache : bool = True) -> list:
-        if self.cache != None and usecache:
-            print("Using cache...")
-            return self.cache
-        length = len(self.gates[0])
 
-        state = self.qbits[0]
-        for each in range(1, self.nqbits): state = qtensor(state, self.qbits[each])
-
-        for each in range(0, length):
-            gates = []
-            for i in range(self.nqbits): 
-                if verbose: print("ngates", len(gates))
-                if self.gates[i][each].gateid == 'cnot': 
-                    if verbose: print('activated')
-                    del gates[-1]
-                if i == 0: gates.append(self.gates[i][each])
-                elif self.gates[i - 1][each].gateid != 'fcnot': gates.append(self.gates[i][each])
-            if verbose: print(len(gates))
-            tens = gates[0]
-            # printreal(tens)
-            for each in range(1, len(gates)): 
-                if verbose: printreal(tens)
-                tens = mtensor(tens, gates[each])
-            if verbose: printreal(tens)
-            state = tens ** state
-            if verbose: print(state)
-        
-        if self.bchange is not None: state = self.bchange.get(self.nqbits) ** state
-        
-        self.cache = state
-        return state
-
-    def measure(self, verbose : bool = False, usecache : bool = True) -> list:
+    def measure(self) -> list:
+        """
+        Performs and returns the result of a single measurement.
+        """
         state = self.qbits[0]
         for each in range(1, self.nqbits): state = qtensor(state, self.qbits[each])
         state = (self.programmat * state).getlist()
 
         return MEASURE(state)
     
-    def run(self, shots : int = 1600, verbose : bool = False, binary : bool  = True, graph : bool = False, usecache : bool = True, terminal : bool = False, legacy : bool = False) -> None:
-        if legacy: state = self.getstate(verbose = verbose, usecache=usecache)
-        else:
-            state = self.qbits[0]
-            for each in range(1, self.nqbits): state = qtensor(state, self.qbits[each])
+    def run(self, shots : int = 1600, binary : bool  = True, graph : bool = False, terminal : bool = False) -> None:
+        """
+        Runs and records the outcome of the qprogram.
 
-            state = (self.programmat * state).getlist()
+        `shots` -> No. of times the measurement of the final state of the system is repeated.
+        `binary` -> Format of the numeric states displayed in the output.
+        `graph` -> If set to True, the output of the qprogram is plotted graphically. May use matplotlib if installed.
+        `terminal` -> Force all the output (including graphs) to be displayed only on the terminal window.
+        """
+        
+        state = self.qbits[0]
+        for each in range(1, self.nqbits): state = qtensor(state, self.qbits[each])
+
+        state = (self.programmat * state).getlist()
 
         run(shots, state, binary=binary, graph=graph, terminal=terminal, name = self.name)
+    
+    def getblock(self, blockid : str = None) -> Block:
+        """
+        Converts the program to a block and returns it.
+
+        If a block id is not provided, program name is used in its stead.
+        """
+        self.compile(showcompilationresult=False)
+        block = Block(self.nqbits, self.name)
+        block.programmat = deepcopy(self.programmat)
+        return block
 
 
 class Block(qprogram):
-    def __init__(self, nqbits: int, blockid: str, custom: list = None, bchange: basischange = None) -> None:
+    """
+    A group of gates' arrangement.
+
+    Blocks can be used in exactly the same way as normal gates, and can contain Blocks themselves.
+
+    A qprogram containing Blocks recursively and automatically compiles all the child blocks, hence blocks can be changed dynamically.
+    """
+    def __init__(self, nqbits: int, blockid: str) -> None:
         self.gateid = blockid
-        super().__init__(nqbits, custom, bchange, f"{blockid}_block")
+        super().__init__(nqbits, name = f"{blockid}_block")
     
     def __repr__(self) -> str:
         if self.repr is None : self.compile(showcompilationresult = False)
         return self.repr
 
     def compile(self, verbose: bool = False, showcompilationresult: bool = True):
+        """
+        Compiles the block (not necessary to be done by the user).
+        """
         super().compile(verbose, showcompilationresult)
         self.span = self.programmat.span
         self.shape = self.programmat.shape
@@ -865,12 +791,19 @@ class Block(qprogram):
         self.rows = self.programmat.rows
     
     def getmat(self) -> Matrix:
-        if self.programmat is None: self.compile()
+        """
+        Returns the plain corresponding to this Block.
+        """
+        self.compile(showcompilationresult = False)
         blockmat = self.programmat
         blockmat.gateid = self.blockid
         return blockmat
 
 def controlledU(u : Matrix) -> Matrix:
+    """
+    Returns the controlled version of gate U, assuming the top qubit (q0) is the control qubit and U is applied on the
+    subsequent set of qubits.
+    """
     cu = Matrix(pow(2, u.span + 1), id = f"C_{u.gateid}")
     
     start = pow(2, u.span)
