+++ refactoring.py
@@ -1,5 +1,7 @@
 from quantum import *
 
-ch = controlledU(cnot0)
+ch = controlledU(CNOT0)
 print(ch.shape)
 print(ch)
+
+comp.getcomplex()
